pyproject.toml
[tool.poetry]
name = "xrandr-monitor-split"
version = "0.1.1"
description = ""
authors = ["Yehonatan Elentok <elyoni@gmail.com>"]
readme = "README.adoc"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"
typer = "^0.12.3"
rich = "^13.7.1"

[tool.poetry.scripts]
xms = 'src.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
